players list was moved to single component
