+++ components/gamePage/GameLogPanel.vue
