+++ pages/game.vue
@@ -1,17 +1,7 @@
 <template lang="pug">
   section
     v-row.flex-nowrap.ma-8
-      .players__wrap(v-show="gameLoaded") 
-        .players__list-toggle(@click="togglePlayerList()")
-           v-icon(class="players__list-toggle-icon mdi" color="white" size="40"  :class="[showPlayerList ? 'mdi-chevron-left'  : 'mdi-chevron-right']")
-        .d-flex.flex-column.mr-5.players__list
-          template(v-for="(player, index) in players")
-            .d-flex
-              userChip(
-                :player="player"
-                :number="index + 1"
-                :isOpened="showPlayerList"
-              )
+      PlayersList(v-show="gameLoaded" v-bind:players="players" @togglePlayerList="showPlayerList = !showPlayerList") 
       .player-zone__wrap(:class="{'collapsed': showPlayerList}")
         v-col.pa-0.player__section(cols="auto" style="overflow: auto; height:500px; widht: 50%" )
           .player__wrap(@click="playPause()" @mouseover="showActionsState = true" @mouseleave="showActionsState = false" ref="playerWrap")
@@ -55,8 +45,8 @@
 </template>
 
 <script>
-import userChip from '@/components/UserChip'
-import GameLogPanel from '@/components/GameLogPanel'
+import GameLogPanel from '@/components/gamePage/GameLogPanel'
+import PlayersList from '@/components/gamePage/PlayersList'
 var player = null
 export default {
   data: () => ({
@@ -64,12 +54,12 @@ export default {
     theme: 1,
     players: [],
     isGameEnd: false,
-    showPlayerList: false,
-    gameLoaded: false
+    gameLoaded: false,
+    showPlayerList: false
   }),
   components: {
-    userChip,
-    GameLogPanel
+    GameLogPanel,
+    PlayersList
   },
   mounted() {
     const base = 'https://storage.googleapis.com/anthive-prod-games/'
@@ -105,9 +95,6 @@ export default {
     },
     showActions() {
       this.showActionsState = !this.showActionsState
-    },
-    togglePlayerList() {
-      this.showPlayerList = !this.showPlayerList
     },
     async copyToClipboard() {
       try {
@@ -146,51 +133,6 @@ export default {
 #player {
   background-repeat: repeat;
   background-color: #388e3c;
-}
-
-.players__wrap {
-  position: relative;
-}
-
-.players__list {
-  height: 987px;
-  overflow-x: hidden;
-  overflow-y: auto;
-  position: relative;
-  background-color: #1a1a1a;
-}
-
-.players__list-toggle {
-  position: absolute;
-  width: 40px;
-  height: 60px;
-  right: -20px;
-  z-index: 1;
-  background: rgba(25, 25, 25, 0.8);
-  border-radius: 0px 8px 8px 0px;
-  cursor: pointer;
-}
-
-.players__list-toggle-icon {
-  font-size: 40px;
-  margin: 10px;
-  margin-left: 0px;
-  color: white;
-}
-
-.players__list::-webkit-scrollbar-track {
-  box-shadow: inset 0 0 6px rgba(0, 0, 0, 0.3);
-  background-color: #1a1a1a;
-}
-
-.players__list::-webkit-scrollbar {
-  width: 5px;
-  background-color: #f5f5f5;
-}
-
-.players__list::-webkit-scrollbar-thumb {
-  background-color: #403f3f;
-  border: none;
 }
 
 .game__vs-separator {
